extracted lowering match logic to seperate file
